add test and Gui_utils python files
